LayoutTests/sputnik/Conformance/12_Statement/12.1_Block/S12.1_A4_T1-expected.txt

S12.1_A4_T1

PASS TypeError: Object is not a function (evaluating '{__func}()')

TEST COMPLETE
